backend/requirements.txt
loguru==0.7.2
httpx==0.28.1
python-dotenv==1.0.1
firecrawl-py==2.16.5
google-generativeai==0.8.3
openai==1.12.0
sendgrid==6.11.0
mailersend==0.5.0
fastapi==0.112.1
uvicorn==0.30.6
